requirements.txt
langchain==0.0.284
chromadb==0.4.15
python-dotenv==1.0.0
fastapi==0.104.0
gunicorn==21.2.0
openai==0.28.1
tiktoken==0.5.1
urllib3==1.25.11
python-multipart==0.0.6
websockets==11.0.3